java to json improved
