+++ src/main/java/com/qa/accountapp/accountapp/App.java
@@ -14,45 +14,28 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import accounts.Account;
 import accounts.AccountService;
+import converters.JavaToJSONConverter;
 import converters.JavaToXMLConverter;
 
 public class App 
 {
     public static void main( String[] args )
     {
-    	
         System.out.println( "Hello world to the standard out" );
+        
         App app = new App();
         AccountService accServ = new AccountService();
         JavaToXMLConverter jxml = new JavaToXMLConverter();
+        JavaToJSONConverter jjson = new JavaToJSONConverter();
+        
         app.populateHash(accServ);
-        app.convertHashToJson(accServ);
+        
+        jjson.convertHashToJson(accServ);
         
         jxml.convertHashToXML(accServ, accServ);
     }
     
-    public void convertHashToJson(AccountService accServ) {
-        HashMap<Integer,Account> hashToConvert = accServ.getHash();
-        
-        ObjectMapper mapper = new ObjectMapper();
-        
-        try {
-        	mapper.writeValue(new File("C:\\Users\\Admin\\eclipse-workspace\\generated\\hashmap.json"), hashToConvert);
-        	
-        	String jsonInStringPls = mapper.writeValueAsString(hashToConvert);
-        	System.out.println(jsonInStringPls);
-        	
-        	String jsonInString = mapper.writerWithDefaultPrettyPrinter().writeValueAsString(hashToConvert);
-        	System.out.println(jsonInString);
-   
-        } catch (JsonGenerationException e) {
-        	e.printStackTrace();
-        } catch (JsonMappingException e) {
-        	e.printStackTrace();
-        } catch (IOException e) {
-        	e.printStackTrace();
-        }
-    }
+
     
     public void populateHash(AccountService accServ) {
     	Account account1 = new Account("Ekaterina", "Genova", 941115);
